refactor around ContentAddressableStorage

+++ common/src/main/scala/cloudcmd/common/adapters/DirectS3Adapter.scala
@@ -1,16 +1,14 @@
 package cloudcmd.common.adapters
 
-import cloudcmd.common.util.{CryptoUtil, StreamUtil}
-import cloudcmd.common.{FileUtil, FileChannelBuffer, UriUtil}
-import org.jboss.netty.buffer.ChannelBuffer
+import cloudcmd.common.util.{JsonUtil, CryptoUtil, StreamUtil}
+import cloudcmd.common.{FileMetaData, BlockContext, FileUtil, UriUtil}
 import org.jets3t.service.impl.rest.httpclient.RestS3Service
 import org.jets3t.service.model.S3Object
 import org.jets3t.service.security.AWSCredentials
-import java.io.ByteArrayInputStream
-import java.io.FileInputStream
-import java.io.InputStream
+import java.io.{File, ByteArrayInputStream, FileInputStream, InputStream}
 import java.net.URI
 import java.nio.channels.Channels
+import collection.mutable
 
 //     "s3://<aws id>@<bucket>?tier=2&tags=s3&secret=<aws secret>"
 
@@ -19,8 +17,7 @@ class DirectS3Adapter extends Adapter {
   private var _bucketName: String = null
   private var _s3Service: RestS3Service = null
 
-  override
-  def init(configDir: String, tier: Int, adapterType: String, tags: Set[String], uri: URI) {
+  override def init(configDir: String, tier: Int, adapterType: String, tags: Set[String], uri: URI) {
     super.init(configDir, tier, adapterType, tags, uri)
     val (awsKey, awsSecret, awsBucketName) = parseAwsInfo(uri)
     val creds = new AWSCredentials(awsKey, awsSecret)
@@ -33,6 +30,8 @@ class DirectS3Adapter extends Adapter {
     }
   }
 
+  def shutdown {}
+
   def parseAwsInfo(adapterUri: URI): (String, String, String) = {
     val parts = adapterUri.getAuthority.split("@")
     if (parts.length != 2) throw new IllegalArgumentException("authority format: awsKey@bucketname")
@@ -41,35 +40,33 @@ class DirectS3Adapter extends Adapter {
     (parts(0), queryParams.get("secret"), parts(1))
   }
 
-  override
   def refreshCache {}
 
-  override
-  def contains(hash: String): Boolean = _s3Service.isObjectInBucket(_bucketName, hash)
-
-  override
-  def shutdown {}
-
-  override
-  def remove(hash: String): Boolean = {
-    _s3Service.deleteObject(_bucketName, hash)
-    true
+  def containsAll(ctxs: Set[BlockContext]): Map[BlockContext, Boolean] = {
+    Map() ++ ctxs.par.flatMap{ctx =>
+      Map(ctx -> _s3Service.isObjectInBucket(_bucketName, ctx.hash))
+    }
   }
 
-  override
-  def verify(hash: String): Boolean = {
-    true
+  def removeAll(ctxs: Set[BlockContext]): Map[BlockContext, Boolean] = {
+    Map() ++ ctxs.par.flatMap{ctx =>
+      _s3Service.deleteObject(_bucketName, ctx.hash)
+      Map(ctx -> true)
+    }
   }
 
-  override
-  def store(is: InputStream, hash: String) {
+  def ensureAll(ctxs: Set[BlockContext], blockLevelCheck: Boolean) : Map[BlockContext, Boolean] = {
+    Map() ++ ctxs.par.flatMap(ctxs => Map(ctxs -> true))
+  }
+
+  def store(ctx: BlockContext, is: InputStream) {
     if (is.isInstanceOf[ByteArrayInputStream]) {
       val buffer = is.asInstanceOf[ByteArrayInputStream]
       try {
         buffer.mark(0)
         val md5Hash = CryptoUtil.computeMD5Hash(Channels.newChannel(buffer))
         buffer.reset
-        store(buffer, hash, md5Hash, buffer.available)
+        store(ctx, buffer, md5Hash, buffer.available)
       } finally {
         buffer.close()
       }
@@ -80,16 +77,17 @@ class DirectS3Adapter extends Adapter {
         buffer.mark(0)
         val md5Hash = CryptoUtil.computeMD5Hash(buffer.getChannel)
         buffer.reset
-        store(buffer, hash, md5Hash, buffer.available)
+        store(ctx, buffer, md5Hash, buffer.available)
       } finally {
         buffer.close()
       }
     }
     else {
       val (hash, tmpFile) = StreamUtil.spoolStream(is)
+      if (hash != ctx.hash) throw new RuntimeException("retrieved data hash %s not equal to expected %s".format(hash, ctx.hash))
       val fis = new FileInputStream(tmpFile)
       try {
-        store(fis, hash)
+        store(ctx, fis)
       } finally {
         fis.close()
         FileUtil.delete(tmpFile)
@@ -97,9 +95,9 @@ class DirectS3Adapter extends Adapter {
     }
   }
 
-  private def store(data: InputStream, hash: String, md5Digest: Array[Byte], length: Long) {
-    if (contains(hash)) return
-    val s3Object: S3Object = new S3Object(hash)
+  private def store(ctx: BlockContext, data: InputStream, md5Digest: Array[Byte], length: Long) {
+    if (contains(ctx)) return
+    val s3Object: S3Object = new S3Object(ctx.hash)
     s3Object.setDataInputStream(data)
     s3Object.setContentLength(length)
     s3Object.setMd5Hash(md5Digest)
@@ -107,25 +105,36 @@ class DirectS3Adapter extends Adapter {
     _s3Service.putObject(_bucketName, s3Object)
   }
 
-  override
-  def load(hash: String): InputStream = {
-    if (!contains(hash)) throw new DataNotFoundException(hash)
-    _s3Service.getObject(_bucketName, hash).getDataInputStream
+  def load(ctx: BlockContext): InputStream = {
+    if (!contains(ctx)) throw new DataNotFoundException(ctx.hash)
+    _s3Service.getObject(_bucketName, ctx.hash).getDataInputStream
   }
 
-  override
-  def loadChannel(hash: String): ChannelBuffer = {
-    if (!contains(hash)) throw new DataNotFoundException(hash)
-    val s3Object: S3Object = _s3Service.getObject(_bucketName, hash)
-    val length = s3Object.getContentLength.toInt
-    new FileChannelBuffer(s3Object.getDataInputStream, length)
-  }
+  def describe: Set[BlockContext] = {
+    val hashes = describeHashes
 
-  override
-  def describe: Set[String] = {
-    Set() ++ _s3Service.listObjects(_bucketName).flatMap {
-      s3Object =>
-        Set(s3Object.getKey)
+    val ctxs = new mutable.HashSet[BlockContext] with mutable.SynchronizedSet[BlockContext]
+    hashes.par.filter(h => h.endsWith(".meta")).par.foreach{ hash =>
+      val fis = _s3Service.getObject(_bucketName, hash).getDataInputStream
+      try {
+        val fmd = FileMetaData.fromJson(hash, JsonUtil.loadJson(fis))
+        ctxs.add(fmd.createBlockContext)
+        val blockHashes = fmd.getBlockHashes
+        (0 until blockHashes.length()).foreach{i =>
+          val blockHash = blockHashes.getString(i)
+          if (hashes.contains(blockHash)) {
+            ctxs.add(fmd.createBlockContext(blockHash))
+          }
+        }
+      } finally {
+        FileUtil.SafeClose(fis)
+      }
     }
+    ctxs.toSet
+  }
+
+
+  def describeHashes: Set[String] = {
+    Set() ++ _s3Service.listObjects(_bucketName).par.map(s3Object => s3Object.getKey)
   }
 }
